fix typescript errors in storybook files
