+++ .storybook/.babelrc
@@ -20,7 +20,8 @@
   ],
   "presets": [
     "@babel/preset-env",
+    "@babel/preset-react",
     "@babel/preset-typescript",
-    "@babel/preset-react"
+		"@babel/preset-flow"
   ]
 }
