--- src/decorators/LowZoomDecorator.js
@@ -1,33 +1,0 @@
-import React, { useMemo } from 'react';
-
-import { MapComponentsProvider } from '../index';
-import MapLibreMap from '../components/MapLibreMap/MapLibreMap';
-import './style.css';
-import { ThemeProvider as MUIThemeProvider } from '@mui/material/styles';
-import getTheme from '../ui_components/MapcomponentsTheme';
-
-const decorators = [
-	(Story, context) => {
-		const theme = useMemo(() => getTheme(context?.globals?.theme), [context?.globals?.theme]);
-
-		return (
-			<div className="fullscreen_map">
-				<MapComponentsProvider>
-					<MUIThemeProvider theme={theme}>
-						<Story />
-						<MapLibreMap
-							options={{
-								zoom: 3,
-								style: 'https://wms.wheregroup.com/tileserver/style/osm-bright.json',
-								center: [4.542400, 39.44518],
-							}}
-							mapId="map_1"
-						/>
-					</MUIThemeProvider>
-				</MapComponentsProvider>
-			</div>
-		);
-	},
-];
-
-export default decorators;
